Update server.js
Some devs may not be familiar with the IPv6 [::] output on local host.  (see https://stackoverflow.com/questions/33853695/node-js-server-address-address-returns )
Added 'localhost' for this example, so the console log will printout the current I[ of localhost.  Make it easier for a user to cmd  click on the console log to launch the application and lowers the friction to get started.